+++ web_poet/__init__.py
@@ -1,2 +1,2 @@
-from .pages import WebPage, ItemPage, ItemWebPage, Injectable
+from .pages import WebPage, JsonPage, ItemPage, ItemWebPage, ItemJsonPage, Injectable
 from .page_inputs import ResponseData